refactor(xy_chart): create basic folder org for chart types
