+++ src/specs/index.ts
@@ -1,8 +1,3 @@
-export { Axis } from './axis';
-export { LineAnnotation } from './line_annotation';
-export { RectAnnotation } from './rect_annotation';
-export { LineSeries } from './line_series';
-export { BarSeries } from './bar_series';
-export { HistogramBarSeries } from './histogram_bar_series';
-export { AreaSeries } from './area_series';
 export { Settings } from './settings';
+// XY chart specs
+export * from '../types/xy_chart/specs';
