+++ src/types/xy_chart/specs/axis.tsx
@@ -1,8 +1,8 @@
 import { inject } from 'mobx-react';
 import { PureComponent } from 'react';
-import { AxisSpec as AxisSpecType, Position } from '../lib/series/specs';
-import { getGroupId } from '../utils/ids';
-import { SpecProps } from './specs_parser';
+import { AxisSpec as AxisSpecType, Position } from '../../../lib/series/specs';
+import { getGroupId } from '../../../utils/ids';
+import { SpecProps } from '../../../specs/specs_parser';
 
 type AxisSpecProps = SpecProps & AxisSpecType;
