+++ src/types/xy_chart/specs/bar_series.tsx
@@ -1,14 +1,14 @@
 import { inject } from 'mobx-react';
 import { PureComponent } from 'react';
-import { HistogramBarSeriesSpec } from '../lib/series/specs';
-import { getGroupId } from '../utils/ids';
-import { ScaleType } from '../utils/scales/scales';
-import { SpecProps } from './specs_parser';
+import { BarSeriesSpec } from '../../../lib/series/specs';
+import { getGroupId } from '../../../utils/ids';
+import { ScaleType } from '../../../utils/scales/scales';
+import { SpecProps } from '../../../specs/specs_parser';
 
-type HistogramBarSpecProps = SpecProps & HistogramBarSeriesSpec;
+type BarSpecProps = SpecProps & BarSeriesSpec;
 
-export class HistogramBarSeriesSpecComponent extends PureComponent<HistogramBarSpecProps> {
-  static defaultProps: Partial<HistogramBarSpecProps> = {
+export class BarSeriesSpecComponent extends PureComponent<BarSpecProps> {
+  static defaultProps: Partial<BarSpecProps> = {
     seriesType: 'bar',
     groupId: getGroupId('__global__'),
     xScaleType: ScaleType.Ordinal,
@@ -17,13 +17,14 @@ export class HistogramBarSeriesSpecComponent extends PureComponent<HistogramBarS
     yAccessors: ['y'],
     yScaleToDataExtent: false,
     hideInLegend: false,
-    enableHistogramMode: true,
+    enableHistogramMode: false,
+    stackAsPercentage: false,
   };
   componentDidMount() {
     const { chartStore, children, ...config } = this.props;
     chartStore!.addSeriesSpec({ ...config });
   }
-  componentDidUpdate(prevProps: HistogramBarSpecProps) {
+  componentDidUpdate(prevProps: BarSpecProps) {
     const { chartStore, children, ...config } = this.props;
     chartStore!.addSeriesSpec({ ...config });
     if (prevProps.id !== this.props.id) {
@@ -39,4 +40,4 @@ export class HistogramBarSeriesSpecComponent extends PureComponent<HistogramBarS
   }
 }
 
-export const HistogramBarSeries = inject('chartStore')(HistogramBarSeriesSpecComponent);
+export const BarSeries = inject('chartStore')(BarSeriesSpecComponent);
