+++ src/types/xy_chart/specs/index.ts
@@ -1,0 +1,7 @@
+export { Axis } from './axis';
+export { LineAnnotation } from './line_annotation';
+export { RectAnnotation } from './rect_annotation';
+export { LineSeries } from './line_series';
+export { BarSeries } from './bar_series';
+export { HistogramBarSeries } from './histogram_bar_series';
+export { AreaSeries } from './area_series';
